requirements.txt
Flask==1.0.2
Flask-Login==0.4.1
Flask-Migrate==2.2.1
Flask-RESTful==0.3.7
Flask-SQLAlchemy==2.3.2
Flask-WTF==0.14.2
Jinja2==2.10
python-dotenv==0.9.1
WTForms==2.2.1
mysqlclient
psutil
